+++ internal/resolver/testdata/helm/repository/kubernetes-charts-index.yaml
