+++ pkg/downloader/manager.go
@@ -30,14 +30,14 @@ import (
 	"github.com/pkg/errors"
 	"sigs.k8s.io/yaml"
 
+	"helm.sh/helm/internal/resolver"
+	"helm.sh/helm/internal/urlutil"
 	"helm.sh/helm/pkg/chart"
 	"helm.sh/helm/pkg/chart/loader"
 	"helm.sh/helm/pkg/chartutil"
 	"helm.sh/helm/pkg/getter"
 	"helm.sh/helm/pkg/helmpath"
 	"helm.sh/helm/pkg/repo"
-	"helm.sh/helm/pkg/resolver"
-	"helm.sh/helm/pkg/urlutil"
 )
 
 // Manager handles the lifecycle of fetching, resolving, and storing dependencies.
